+++ lint_test/avoid_using_api/external_source/lib/external_source.dart
@@ -22,3 +22,9 @@ const test2 = 'Hello World';
 void test() {}
 
 int banned = 5;
+
+extension BannedExtension on int {
+  int banned() => this + 10;
+
+  int get bannedGetter => 10;
+}