+++ lint_test/avoid_using_api/identifier_extension_source_ban/lib/identifier_extension_source_ban_test.dart
@@ -1,0 +1,11 @@
+import 'package:external_source/external_source.dart';
+
+void extensionIdentifierBanTesting() {
+  const int a = 10;
+
+  // expect_lint: avoid_using_api
+  a.banned();
+
+  // expect_lint: avoid_using_api
+  a.bannedGetter;
+}